+++ modules/Link.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import warning from './routerWarning'
 import invariant from 'invariant'
 import { routerShape } from './PropTypes'
+import { ContextSubscriber } from './ContextUtils'
 
 const { bool, object, string, func, oneOfType } = React.PropTypes
 
@@ -22,12 +22,8 @@ function isEmptyObject(object) {
   return true
 }
 
-function createLocationDescriptor(to, { query, hash, state }) {
-  if (query || hash || state) {
-    return { pathname: to, query, hash, state }
-  }
-
-  return to
+function resolveToLocation(to, router) {
+  return typeof to === 'function' ? to(router.location) : to
 }
 
 /**
@@ -50,12 +46,14 @@ function createLocationDescriptor(to, { query, hash, state }) {
  */
 const Link = React.createClass({
 
+  mixins: [ ContextSubscriber('router') ],
+
   contextTypes: {
     router: routerShape
   },
 
   propTypes: {
-    to: oneOfType([ string, object ]),
+    to: oneOfType([ string, object, func ]),
     query: object,
     hash: string,
     state: object,
@@ -80,8 +78,9 @@ const Link = React.createClass({
     if (event.defaultPrevented)
       return
 
+    const { router } = this.context
     invariant(
-      this.context.router,
+      router,
       '<Link>s rendered outside of a router context cannot navigate.'
     )
 
@@ -95,31 +94,24 @@ const Link = React.createClass({
 
     event.preventDefault()
 
-    const { to, query, hash, state } = this.props
-    const location = createLocationDescriptor(to, { query, hash, state })
-
-    this.context.router.push(location)
+    router.push(resolveToLocation(this.props.to, router))
   },
 
   render() {
-    const { to, query, hash, state, activeClassName, activeStyle, onlyActiveOnIndex, ...props } = this.props
-    warning(
-      !(query || hash || state),
-      'the `query`, `hash`, and `state` props on `<Link>` are deprecated, use `<Link to={{ pathname, query, hash, state }}/>. http://tiny.cc/router-isActivedeprecated'
-    )
+    const { to, activeClassName, activeStyle, onlyActiveOnIndex, ...props } = this.props
 
-    // Ignore if rendered outside the context of router, simplifies unit testing.
+    // Ignore if rendered outside the context of router to simplify unit testing.
     const { router } = this.context
 
     if (router) {
       // If user does not specify a `to` prop, return an empty anchor tag.
       if (to == null) { return <a {...props} /> }
 
-      const location = createLocationDescriptor(to, { query, hash, state })
-      props.href = router.createHref(location)
+      const toLocation = resolveToLocation(to, router)
+      props.href = router.createHref(toLocation)
 
       if (activeClassName || (activeStyle != null && !isEmptyObject(activeStyle))) {
-        if (router.isActive(location, onlyActiveOnIndex)) {
+        if (router.isActive(toLocation, onlyActiveOnIndex)) {
           if (activeClassName) {
             if (props.className) {
               props.className += ` ${activeClassName}`