+++ modules/PromiseUtils.js
@@ -1,0 +1,3 @@
+export function isPromise(obj) {
+  return obj && typeof obj.then === 'function'
+}