+++ modules/Router.js
@@ -1,5 +1,3 @@
-import createHashHistory from 'history/lib/createHashHistory'
-import useQueries from 'history/lib/useQueries'
 import invariant from 'invariant'
 import React from 'react'
 
@@ -7,18 +5,8 @@ import createTransitionManager from './createTransitionManager'
 import { routes } from './InternalPropTypes'
 import RouterContext from './RouterContext'
 import { createRoutes } from './RouteUtils'
-import { createRouterObject, createRoutingHistory } from './RouterUtils'
+import { createRouterObject, assignRouterState } from './RouterUtils'
 import warning from './routerWarning'
-
-function isDeprecatedHistory(history) {
-  return !history || !history.__v2_compatible__
-}
-
-/* istanbul ignore next: sanity check */
-function isUnsupportedHistory(history) {
-  // v3 histories expose getCurrentLocation, but aren't currently supported.
-  return history && history.getCurrentLocation
-}
 
 const { func, object } = React.PropTypes
 
@@ -37,10 +25,6 @@ const Router = React.createClass({
     createElement: func,
     onError: func,
     onUpdate: func,
-
-    // Deprecated:
-    parseQueryString: func,
-    stringifyQuery: func,
 
     // PRIVATE: For client-side rehydration of server match.
     matchContext: object
@@ -72,71 +56,52 @@ const Router = React.createClass({
     }
   },
 
-  componentWillMount() {
-    const { parseQueryString, stringifyQuery } = this.props
-    warning(
-      !(parseQueryString || stringifyQuery),
-      '`parseQueryString` and `stringifyQuery` are deprecated. Please create a custom history. http://tiny.cc/router-customquerystring'
-    )
-
-    const { history, transitionManager, router } = this.createRouterObjects()
-
-    this._unlisten = transitionManager.listen((error, state) => {
-      if (error) {
-        this.handleError(error)
-      } else {
-        this.setState(state, this.props.onUpdate)
-      }
-    })
-
-    this.history = history
-    this.router = router
-  },
-
-  createRouterObjects() {
+  createRouterObject(state) {
     const { matchContext } = this.props
     if (matchContext) {
-      return matchContext
+      return matchContext.router
     }
 
-    let { history } = this.props
+    const { history } = this.props
+    return createRouterObject(history, this.transitionManager, state)
+  },
+
+  createTransitionManager() {
+    const { matchContext } = this.props
+    if (matchContext) {
+      return matchContext.transitionManager
+    }
+
+    const { history } = this.props
     const { routes, children } = this.props
 
     invariant(
-      !isUnsupportedHistory(history),
-      'You have provided a history object created with history v3.x. ' +
-      'This version of React Router is not compatible with v3 history ' +
-      'objects. Please use history v2.x instead.'
+      history.getCurrentLocation,
+      'You have provided a history object created with history v2.x or ' +
+      'earlier. This version of React Router is only compatible with v3 ' +
+      'history objects. Please upgrade to history v3.x.'
     )
 
-    if (isDeprecatedHistory(history)) {
-      history = this.wrapDeprecatedHistory(history)
-    }
-
-    const transitionManager = createTransitionManager(
-      history, createRoutes(routes || children)
+    return createTransitionManager(
+      history,
+      createRoutes(routes || children)
     )
-    const router = createRouterObject(history, transitionManager)
-    const routingHistory = createRoutingHistory(history, transitionManager)
-
-    return { history: routingHistory, transitionManager, router }
   },
 
-  wrapDeprecatedHistory(history) {
-    const { parseQueryString, stringifyQuery } = this.props
+  componentWillMount() {
+    this.transitionManager = this.createTransitionManager()
+    this.router = this.createRouterObject(this.state)
 
-    let createHistory
-    if (history) {
-      warning(false, 'It appears you have provided a deprecated history object to `<Router/>`, please use a history provided by ' +
-                     'React Router with `import { browserHistory } from \'react-router\'` or `import { hashHistory } from \'react-router\'`. ' +
-                     'If you are using a custom history please create it with `useRouterHistory`, see http://tiny.cc/router-usinghistory for details.')
-      createHistory = () => history
-    } else {
-      warning(false, '`Router` no longer defaults the history prop to hash history. Please use the `hashHistory` singleton instead. http://tiny.cc/router-defaulthistory')
-      createHistory = createHashHistory
-    }
-
-    return useQueries(createHistory)({ parseQueryString, stringifyQuery })
+    this._unlisten = this.transitionManager.listen((error, state) => {
+      if (error) {
+        this.handleError(error)
+      } else {
+        // Keep the identity of this.router because of a caveat in ContextUtils:
+        // they only work if the object identity is preserved.
+        assignRouterState(this.router, state)
+        this.setState(state, this.props.onUpdate)
+      }
+    })
   },
 
   /* istanbul ignore next: sanity check */
@@ -171,7 +136,6 @@ const Router = React.createClass({
 
     return render({
       ...props,
-      history: this.history,
       router: this.router,
       location,
       routes,