+++ modules/TransitionUtils.js
@@ -1,5 +1,4 @@
 import { loopAsync } from './AsyncUtils'
-import warning from './routerWarning'
 
 function createTransitionHook(hook, route, asyncArity) {
   return function (...args) {
@@ -38,21 +37,7 @@ function runTransitionHooks(length, iter, callback) {
   }
 
   let redirectInfo
-  function replace(location, deprecatedPathname, deprecatedQuery) {
-    if (deprecatedPathname) {
-      warning(
-        false,
-        '`replaceState(state, pathname, query) is deprecated; use `replace(location)` with a location descriptor instead. http://tiny.cc/router-isActivedeprecated'
-      )
-      redirectInfo = {
-        pathname: deprecatedPathname,
-        query: deprecatedQuery,
-        state: location
-      }
-
-      return
-    }
-
+  function replace(location) {
     redirectInfo = location
   }