+++ modules/createMemoryHistory.js
@@ -3,12 +3,11 @@ import useBasename from 'history/lib/useBasename'
 import baseCreateMemoryHistory from 'history/lib/createMemoryHistory'
 
 export default function createMemoryHistory(options) {
-  // signatures and type checking differ between `useRoutes` and
+  // signatures and type checking differ between `useQueries` and
   // `createMemoryHistory`, have to create `memoryHistory` first because
   // `useQueries` doesn't understand the signature
   const memoryHistory = baseCreateMemoryHistory(options)
   const createHistory = () => memoryHistory
   const history = useQueries(useBasename(createHistory))(options)
-  history.__v2_compatible__ = true
   return history
 }