+++ modules/index.js
@@ -10,17 +10,10 @@ export IndexRoute from './IndexRoute'
 export Redirect from './Redirect'
 export Route from './Route'
 
-/* mixins */
-export History from './History'
-export Lifecycle from './Lifecycle'
-export RouteContext from './RouteContext'
-
 /* utils */
-export useRoutes from './useRoutes'
 export { createRoutes } from './RouteUtils'
 export RouterContext from './RouterContext'
-export RoutingContext from './RoutingContext'
-export PropTypes, { locationShape, routerShape } from './PropTypes'
+export { locationShape, routerShape } from './PropTypes'
 export match from './match'
 export useRouterHistory from './useRouterHistory'
 export { formatPattern } from './PatternUtils'