+++ modules/match.js
@@ -4,7 +4,7 @@ import invariant from 'invariant'
 import createMemoryHistory from './createMemoryHistory'
 import createTransitionManager from './createTransitionManager'
 import { createRoutes } from './RouteUtils'
-import { createRouterObject, createRoutingHistory } from './RouterUtils'
+import { createRouterObject } from './RouterUtils'
 
 /**
  * A high-level API to be used for server-side rendering.
@@ -27,40 +27,30 @@ function match({ history, routes, location, ...options }, callback) {
     createRoutes(routes)
   )
 
-  let unlisten
-
   if (location) {
     // Allow match({ location: '/the/path', ... })
     location = history.createLocation(location)
   } else {
-    // Pick up the location from the history via synchronous history.listen
-    // call if needed.
-    unlisten = history.listen(historyLocation => {
-      location = historyLocation
-    })
+    location = history.getCurrentLocation()
   }
 
-  const router = createRouterObject(history, transitionManager)
-  history = createRoutingHistory(history, transitionManager)
+  transitionManager.match(location, (error, redirectLocation, nextState) => {
+    let renderProps
 
-  transitionManager.match(location, function (error, redirectLocation, nextState) {
+    if (nextState) {
+      const router = createRouterObject(history, transitionManager, nextState)
+      renderProps = {
+        ...nextState,
+        router,
+        matchContext: { transitionManager, router }
+      }
+    }
+
     callback(
       error,
-      redirectLocation && router.createLocation(redirectLocation, REPLACE),
-      nextState && {
-        ...nextState,
-        history,
-        router,
-        matchContext: { history, transitionManager, router }
-      }
+      redirectLocation && history.createLocation(redirectLocation, REPLACE),
+      renderProps
     )
-
-    // Defer removing the listener to here to prevent DOM histories from having
-    // to unwind DOM event listeners unnecessarily, in case callback renders a
-    // <Router> and attaches another history listener.
-    if (unlisten) {
-      unlisten()
-    }
   })
 }